requirements.txt
Flask>=2.2.0
mysql-connector-python>=8.0.0
python-dotenv>=1.0.0 